Ticketify/requirements.txt
Django>=4.2,<5.0
Pillow>=10.0.0
qrcode[pil]>=7.4.2
python-dateutil>=2.8.2
reportlab>=4.0.0
djangorestframework>=3.14
redis>=4.5.0
django-redis>=5.2.0
